refactor: create fragments folder
