+++ ui/src/views/indexed-nfa/fragments/index.ts
@@ -1,0 +1,3 @@
+export * from './aside.fragment';
+export * from './main.fragment';
+export * from './skeleton.fragment';
